+++ examples/terminal_widget.rs
@@ -78,7 +78,7 @@ fn main() -> turbo_vision::core::error::Result<()> {
     let (width, height) = app.terminal.size();
 
     // Create status line
-    let status_line = StatusLine::new(Rect::new(0, height as i16 - 1, width as i16, height as i16), vec![StatusItem::new("~Alt+X~ Exit", KB_ALT_X, CM_QUIT)]);
+    let status_line = StatusLine::new(Rect::new(0, height - 1, width, height), vec![StatusItem::new("~Alt+X~ Exit", KB_ALT_X, CM_QUIT)]);
     app.set_status_line(status_line);
 
     // Create dialog (matches Borland structure but larger for more content)
