+++ examples/wrong_owner.rs
@@ -17,7 +17,7 @@ fn main() -> turbo_vision::core::error::Result<()> {
     // Status line
     let (width, height) = app.terminal.size();
     let status_line = StatusLine::new(
-        Rect::new(0, height as i16 - 1, width as i16, height as i16),
+        Rect::new(0, height - 1, width, height),
         vec![
             StatusItem::new("~Alt-X~ Exit", KB_ALT_X, CM_QUIT),
             StatusItem::new("~Esc~ Exit", KB_ESC, CM_QUIT),
